+++ src/main/java/cn/xiaosm/plainadmin/controller/UserController.java
@@ -11,21 +11,20 @@
 package cn.xiaosm.plainadmin.controller;
 
 import cn.xiaosm.plainadmin.config.security.service.TokenService;
-import cn.xiaosm.plainadmin.entity.ResponseEntity;
+import cn.xiaosm.plainadmin.entity.ResponseBody;
 import cn.xiaosm.plainadmin.entity.User;
 import cn.xiaosm.plainadmin.entity.enums.StatusEnum;
 import cn.xiaosm.plainadmin.entity.vo.UserVO;
+import cn.xiaosm.plainadmin.exception.SQLOperateException;
 import cn.xiaosm.plainadmin.service.UserService;
 import cn.xiaosm.plainadmin.utils.ResponseUtils;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.Date;
 import java.util.Set;
 
 /**
@@ -59,13 +58,13 @@ public class UserController {
      * @return
      */
     @GetMapping("/info")
-    public ResponseEntity userInfo(HttpServletRequest request) {
+    public ResponseBody userInfo(HttpServletRequest request) {
         return ResponseUtils.buildSuccess(tokenService.getLoginUser(request));
     }
 
     @GetMapping("")
     @PreAuthorize("hasAuthority('user:query') or hasRole('admin')")
-    public ResponseEntity queryUsers(Page<User> page, UserVO userVO) {
+    public ResponseBody queryUsers(Page<User> page, UserVO userVO) {
         QueryWrapper<User> wrapper = new QueryWrapper<>();
 
         // 不查询已删除的用户
@@ -76,7 +75,7 @@ public class UserController {
 
     @PutMapping
     @PreAuthorize("(hasAuthority('user:add') and hasAuthority('role:query')) or hasRole('admin')")
-    public ResponseEntity saveUser(@RequestBody UserVO userVO) {
+    public ResponseBody saveUser(@RequestBody UserVO userVO) {
         boolean b = userService.addEntity(userVO);
         return b == true ? ResponseUtils.buildSuccess("新增用户信息成功")
                 : ResponseUtils.buildFail("保存失败");
@@ -84,7 +83,8 @@ public class UserController {
 
     @PostMapping
     @PreAuthorize("hasAuthority('user:modify') or hasRole('admin')")
-    public ResponseEntity modifyUser(@RequestBody UserVO userVO) {
+    public ResponseBody modifyUser(@RequestBody UserVO userVO) {
+        if (userVO.getId() == 1) throw new SQLOperateException("系统保留数据，请勿操作");
         boolean b = userService.modifyEntity(userVO);
         return b == true ? ResponseUtils.buildSuccess("修改用户信息成功")
                 : ResponseUtils.buildFail("修改失败");
@@ -92,7 +92,10 @@ public class UserController {
 
     @DeleteMapping
     @PreAuthorize("hasAuthority('user:delete') or hasRole('admin')")
-    public ResponseEntity deleteUsers(@RequestBody Set<Integer> ids) {
+    public ResponseBody deleteUsers(@RequestBody Set<Integer> ids) {
+        if (ids.stream().filter(el -> el == 1).count() == 1) {
+            throw new SQLOperateException("系统保留数据，请勿操作");
+        }
         int b = userService.removeByIds(ids);
         return ResponseUtils.buildSuccess("删除用户成功，本次删除" + b + "条用户");
     }
